pyproject.toml
[tool.poetry]
name = "ncku-pastexam-backend"
version = "0.1.0"
description = ""
authors = ["Vincent550102 <5020559@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
mysql-connector-python = "^8.1.0"
fastapi = {extras = ["all"], version = "^0.103.2"}
google-api-python-client = "^2.101.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
